+++ packages/react/src/components/WizardProvider.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import type { FC, ReactNode } from "react";
 import type { WalletError } from "@wizard-ui/core";
 
-import { CWClientProvider } from "./CWClientProvider";
 import { WalletModalProvider } from "./WalletModalProvider";
 import { WalletProvider } from "./WalletProvider";
 
@@ -18,14 +17,11 @@ export interface WizardProviderProps {
 
 export const WizardProvider: FC<WizardProviderProps> = ({
   children,
-  endpoint,
   ...rest
 }) => {
   return (
-    <CWClientProvider endpoint={endpoint}>
-      <WalletProvider {...rest}>
-        <WalletModalProvider>{children}</WalletModalProvider>
-      </WalletProvider>
-    </CWClientProvider>
+    <WalletProvider {...rest}>
+      <WalletModalProvider>{children}</WalletModalProvider>
+    </WalletProvider>
   );
 };