+++ packages/react/src/hooks/index.ts
@@ -1,5 +1,3 @@
-export * from "./useBalance";
-export * from "./useCWClient";
 export * from "./useExecuteContract";
 export * from "./useQueryContract";
 export * from "./useWalletModal";